# docs/swagger.yaml
definitions:
  controller.BindMachineExample:
    properties:
      machine_code:
        example: "23333333333"
        type: string
    type: object
  controller.ExternalLoginRegisterExample:
    properties:
      machine_code:
        example: "23333333333"
        type: string
      password:
        example: qianyi
        type: string
      username:
        example: admin
        type: string
    type: object
  controller.ExternalRegisterExample:
    properties:
      machine_code:
        example: "23333333333"
        type: string
    type: object
  controller.UpdateOrderExample:
    properties:
      cnum:
        example: "1"
        type: string
      log:
        example: 测试log
        type: string
      status:
        example: "1"
        type: string
    type: object
  model.Game:
    properties:
      code:
        description: 游戏编码
        type: string
      id:
        type: integer
      mode_list:
        description: 游戏模式
        items:
          $ref: '#/definitions/model.Mode'
        type: array
      name:
        description: 游戏名称
        type: string
      option:
        description: 游戏选项
        items:
          $ref: '#/definitions/model.Option'
        type: array
      ver:
        description: 版本号
        type: string
    type: object
  model.Log:
    properties:
      created_at:
        type: string
      id:
        type: integer
      num:
        description: 数量
        type: integer
      order_id:
        description: 订单ID
        type: integer
      status:
        description: 状态
        type: integer
      text:
        description: 内容
        type: string
    type: object
  model.Mode:
    properties:
      code:
        description: 模式编码
        type: string
      id:
        type: integer
      name:
        description: 模式名称 （例：安卓，ios）
        type: string
    type: object
  model.Option:
    properties:
      code:
        description: 选项编码
        type: string
      game_id:
        description: 游戏ID
        type: string
      id:
        type: integer
      name:
        description: 选项名称
        type: string
      option_items:
        description: 选项内容
        items:
          $ref: '#/definitions/model.OptionItem'
        type: array
    type: object
  model.OptionItem:
    properties:
      code:
        description: 选项内容编码
        type: string
      id:
        type: integer
      name:
        description: 选项内容名称
        type: string
      option_id:
        description: 选项ID
        type: integer
    type: object
  model.Order:
    properties:
      cnum:
        description: 完成数量
        type: integer
      game:
        $ref: '#/definitions/model.Game'
        description: 游戏详情
      id:
        type: integer
      log:
        description: 日志详情
        items:
          $ref: '#/definitions/model.Log'
        type: array
      machine:
        description: 绑定机器码
        type: string
      mode:
        $ref: '#/definitions/model.Mode'
        description: 模式详情
      name:
        description: 账号用户名
        type: string
      num:
        description: 数量
        type: integer
      option:
        description: 选项
        type: string
      order_id:
        description: 订单id
        type: integer
      password:
        description: 账号密码
        type: string
      status:
        description: 订单状态 0 初始化 1 执行中，2 执行完毕， -1 执行失败，-2 暂停订单
        type: integer
      type:
        description: 类型
        type: string
    type: object
  model.User:
    properties:
      code:
        description: 特征码
        type: string
      id:
        type: integer
      machine:
        description: 机器码
        type: string
      name:
        description: 用户账号
        type: string
      password:
        description: 用户密码
        type: string
      token:
        description: 登录密钥
        type: string
    type: object
  model.UserMachine:
    properties:
      id:
        type: integer
      machine:
        description: 用户机器码
        type: string
      machine_id:
        description: 机器码ID
        type: integer
      tag:
        description: 标签
        type: string
      user_id:
        description: 用户id
        type: integer
    type: object
info:
  contact: {}
  description: 下单API
  title: 下单API
  version: "1.0"
paths:
  /api/v1/external/bind/{code}:
    post:
      consumes:
      - application/json
      description: 绑定特征码和机器码
      parameters:
      - default: PKPV48
        description: 特征码
        in: path
        name: code
        required: true
        type: string
      - description: 机器码
        in: body
        name: machine_code
        required: true
        schema:
          $ref: '#/definitions/controller.BindMachineExample'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                success:
                  type: boolean
              type: object
      summary: 绑定特征码和机器码
      tags:
      - V1接口
  /api/v1/external/order/{code}/ing:
    get:
      consumes:
      - application/json
      description: 获取执行中的订单，如果没有执行中的订单会返回报错
      parameters:
      - default: PKPV48
        description: 特征码
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.Order'
                success:
                  type: boolean
              type: object
      summary: 获取执行中的订单
      tags:
      - V1接口
  /api/v1/external/order/{code}/new:
    get:
      consumes:
      - application/json
      description: 获取一个新的订单，如果有正在执行的订单会返回报错
      parameters:
      - default: PKPV48
        description: 特征码
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.Order'
                success:
                  type: boolean
              type: object
      summary: 获取一个新的订单
      tags:
      - V1接口
  /api/v1/external/order/{code}/update/{order_id}:
    post:
      consumes:
      - application/json
      description: 更新的执行中的订单 <br><br> status 状态（0 初始化 1 执行中，2 执行完毕， -1 执行失败，-2 暂停订单）<br> cnum 完成数量 <br> text 日志内容
      parameters:
      - default: PKPV48
        description: 特征码
        in: path
        name: code
        required: true
        type: string
      - default: 20210619215350
        description: 订单ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: 请求参数
        in: body
        name: body
        schema:
          $ref: '#/definitions/controller.UpdateOrderExample'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.Order'
                success:
                  type: boolean
              type: object
      summary: 更新的执行中的订单
      tags:
      - V1接口
  /api/v1/external/test/{code}:
    post:
      consumes:
      - application/json
      description: 测试post接口
      parameters:
      - default: PKPV48
        description: 特征码
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.User'
                success:
                  type: boolean
              type: object
      summary: 测试post接口
      tags:
      - V1接口
  /api/v2/external/login/register:
    post:
      consumes:
      - application/json
      description: 登录绑定机器码
      parameters:
      - description: 机器码
        in: body
        name: machine_code
        required: true
        schema:
          $ref: '#/definitions/controller.ExternalLoginRegisterExample'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.UserMachine'
                message:
                  type: string
                success:
                  type: boolean
              type: object
      summary: 登录绑定机器码 返回用户机器码
      tags:
      - V2接口
  /api/v2/external/machine/deregister:
    post:
      consumes:
      - application/json
      description: 服务注销(机器) 关闭时候调用
      parameters:
      - description: 机器码
        in: body
        name: machine_code
        required: true
        schema:
          $ref: '#/definitions/controller.ExternalRegisterExample'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                message:
                  type: string
                success:
                  type: boolean
              type: object
      summary: 机器服务注销
      tags:
      - V2接口
  /api/v2/external/machine/health:
    post:
      consumes:
      - application/json
      description: 机器心跳包
      parameters:
      - description: 机器码
        in: body
        name: machine_code
        required: true
        schema:
          $ref: '#/definitions/controller.ExternalRegisterExample'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                message:
                  type: string
                success:
                  type: boolean
              type: object
      summary: 机器心跳包
      tags:
      - V2接口
  /api/v2/external/machine/register:
    post:
      consumes:
      - application/json
      description: 服务发现(机器) 初始化时候调用
      parameters:
      - description: 机器码
        in: body
        name: machine_code
        required: true
        schema:
          $ref: '#/definitions/controller.ExternalRegisterExample'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                message:
                  type: string
                success:
                  type: boolean
              type: object
      summary: 机器服务发现
      tags:
      - V2接口
  /api/v2/external/order/{code}/ing:
    get:
      consumes:
      - application/json
      description: 获取执行中的订单，如果没有执行中的订单会返回报错
      parameters:
      - default: 085ebd3b-0c6d-469e-9cb5-f9971d81d223
        description: 用户机器码
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.Order'
                success:
                  type: boolean
              type: object
      summary: 获取执行中的订单
      tags:
      - V2接口
  /api/v2/external/order/{code}/new:
    get:
      consumes:
      - application/json
      description: 获取一个新的订单，如果有正在执行的订单会返回报错
      parameters:
      - default: 085ebd3b-0c6d-469e-9cb5-f9971d81d223
        description: 用户机器码
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.Order'
                success:
                  type: boolean
              type: object
      summary: 获取一个新的订单
      tags:
      - V2接口
  /api/v2/external/order/{code}/update/{order_id}:
    post:
      consumes:
      - application/json
      description: 更新的执行中的订单 <br><br> status 状态（0 初始化 1 执行中，2 执行完毕， -1 执行失败，-2 暂停订单）<br> cnum 完成数量 <br> text 日志内容
      parameters:
      - default: 085ebd3b-0c6d-469e-9cb5-f9971d81d223
        description: 用户机器码
        in: path
        name: code
        required: true
        type: string
      - default: 20210619215350
        description: 订单ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: 请求参数
        in: body
        name: body
        schema:
          $ref: '#/definitions/controller.UpdateOrderExample'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.Order'
                success:
                  type: boolean
              type: object
      summary: 更新的执行中的订单
      tags:
      - V2接口
swagger: "2.0"